requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1 
Flask-Mail==0.9.1
itsdangerous==2.1.2

# PDF text extraction
PyMuPDF==1.24.9

# NLP matching
spacy==3.8.11
sentence-transformers==3.0.1
torch==2.3.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2

apscheduler